Resolve TFVC build server path to local path
If user selected a server path from the path picker, we will attempt to
use tf's resolvePath command to resolve the server path to a local path.
If resolution isn't possible (tf doesn't have the cmd, the input path
is not a server path, or the server path hasn't been mapped yet), we
will return the input.  This way we could still honor any relative path
the customer has entered themselves.